renamed module to torchmetrics
