+++ luxonis_train/attached_modules/metrics/__init__.py
@@ -1,8 +1,8 @@
 from .base_metric import BaseMetric
-from .common import Accuracy, F1Score, JaccardIndex, Precision, Recall
 from .mean_average_precision import MeanAveragePrecision
 from .mean_average_precision_keypoints import MeanAveragePrecisionKeypoints
 from .object_keypoint_similarity import ObjectKeypointSimilarity
+from .torchmetrics import Accuracy, F1Score, JaccardIndex, Precision, Recall
 
 __all__ = [
     "Accuracy",
